chore: analysis and transform stop importing each other
