+++ packages/svelte/src/compiler/phases/3-transform/client/visitors/VariableDeclaration.js
@@ -5,14 +5,8 @@ import { dev, is_ignored, locate_node } from '../../../../state.js';
 import { extract_paths, save } from '../../../../utils/ast.js';
 import * as b from '#compiler/builders';
 import * as assert from '../../../../utils/assert.js';
-import { get_rune } from '../../../scope.js';
-import {
-	async_thunk,
-	get_prop_source,
-	is_prop_source,
-	is_state_source,
-	should_proxy
-} from '../utils.js';
+import { get_rune, should_proxy } from '../../../scope.js';
+import { async_thunk, get_prop_source, is_prop_source, is_state_source } from '../utils.js';
 import { get_value } from './shared/declarations.js';
 
 /**
