+++ packages/svelte/src/compiler/utils/check_graph_for_cycles.js
